+++ src/lib/components/workflow-actions.svelte
@@ -12,7 +12,7 @@
   import SplitButton from '$lib/holocene/split-button.svelte';
   import MenuItem from '$lib/holocene/primitives/menu/menu-item.svelte';
   import { cancelWorkflow } from '$lib/services/workflow-service';
-  import { toaster } from '$holocene/stores/toaster';
+  import { toaster } from '$lib/stores/toaster';
   import Input from '$lib/holocene/input/input.svelte';
   import FeatureGuard from './feature-guard.svelte';
   import Button from '$lib/holocene/button.svelte';
